Tache B8 - Ajout bon typage du retour de demande de proba par capital

+++ headers/interface.h
@@ -82,7 +82,7 @@ int demander_personnage(joueuse j1);
  * \brief Demande à une joueuse quelle probabilité par capital elle souhaite avoir entre 0.1 et 1
  * \return la valeur de probabilité par capital que la joueuse souhaite avoir, 0 sinon
 */
-int demander_proba_par_capital();
+float demander_proba_par_capital();
 
 /**
  * \brief Affiche dans la console un message quand le jeu est fini
